Revert "Merge remote-tracking branch 'SS14/master' into upstream"
This reverts commit 935ae31b30b2529228dea8ae6fef0a1863b212eb.

+++ Content.Server/Singularity/Components/SinguloFoodComponent.cs
@@ -8,6 +8,6 @@ namespace Content.Server.Singularity.Components
     {
         [ViewVariables(VVAccess.ReadWrite)]
         [DataField("energy")]
-        public float Energy { get; set; } = 1f;
+        public int Energy { get; set; } = 1;
     }
 }
